Different Attributes for each ball
Added a scriptable object that stores the color of the ball and other attributes that can modify the balls properties. Changed the name property on ColorSO to an Enum type to make it less error prone for future uses. Included three test scenes that test the ball's different attributes. Red balls are heavier and harder to move, green balls are lighter and easier to move, and blue balls are bouncier.