Additional Progress for MagnetEffect
Created a working version of the Magnet Effect in a sepeare script, MagnetEffect.cs.
This script works using a fixed update. There is a coroutine that could be initiated, but unable to test if this works atm because effects are only created on collision, but magnet effect should take place during instantiation