magnet ball effect (#90)
* Magnet Effect

Created Yellow Ball Scene
Created scriptable object for yellow ball and magnet effect

* Additional updated to the magned SO

Magnet SO is not functional at this time and is still being developed.

* Additional Progress for MagnetEffect

Created a working version of the Magnet Effect in a sepeare script, MagnetEffect.cs.
This script works using a fixed update. There is a coroutine that could be initiated, but unable to test if this works atm because effects are only created on collision, but magnet effect should take place during instantiation

* Magnet Ball v1

Finished Magnet ball script. Currently does not work as a scriptable object and is instead a regular script. Plans to convert to scriptable object in the future.

* magnet effect fixes

---------

Co-authored-by: kallenxcc <braetonbuckley@gmail.com>