Quick Fix for Clock + Changes to Sign Prefab
Made new copy of the clock image with the text removed. This was causing a visual bug when the clock was faded because the text on the image was visible.
Rearranged the Timer prefab so that the Timer Text is not a child of the Timer Background and also set Timer Background to inactive. This was causing a visual discrepancy where the background around the text was more visible compared to the rest of the clock. Seems safe to delete the timer background entirely.
Added missing adjustments to the sign manager prefab and Sign Post Notification prefab so that it will work as intended.