Quick Fixes for Kitchen Scene
Added missing reference values from the Slingshot.cs script in the kitchen scene and disabled the option to pick a challenge and puzzle variant of the levels from the main menu since those scenes arent fixed yet. Intended to be a simple solution made in time for the legion war. Better alternative should be explored afterwards.