Yarn Ball Menu Tweak
Visual updates to some of the examples to use the Gaegu-Regular font on Pluses and Equal signs.

Changed and Future-proofed the button menu layout to allow for more example yarn balls with a scroll down. If you wish to test this out copy and paste the other buttons under the content parent.